# Faster binary distances on Arm NEON & SVE